Extract Edit_Field from Console
