+++ src/something_console.cpp
@@ -4,22 +4,6 @@
 // TODO(#181): console does not support history
 // TODO(#182): console does not support scrolling
 
-Console::Selection Console::get_selection() const
-{
-    Selection result = {};
-
-    if (edit_field_selection_begin <= edit_field_cursor) {
-        result.begin = edit_field_selection_begin;
-        result.end = edit_field_cursor;
-    } else {
-        result.begin = edit_field_cursor + 1;
-        result.end = edit_field_selection_begin + 1;
-    }
-
-    assert(result.begin <= result.end);
-    return result;
-}
-
 void Console::render(SDL_Renderer *renderer, Bitmap_Font *font)
 {
     if (slide_position > 0.0f) {
@@ -44,56 +28,12 @@ void Console::render(SDL_Renderer *renderer, Bitmap_Font *font)
         const auto edit_field_position =
             vec2(prompt_size.x, console_y + (float)(CONSOLE_VISIBLE_ROWS * BITMAP_FONT_CHAR_HEIGHT * CONSOLE_FONT_SIZE));
         font->render(renderer, edit_field_position - vec2(prompt_size.x, 0.0f), vec2(CONSOLE_FONT_SIZE, CONSOLE_FONT_SIZE),
-                     FONT_DEBUG_COLOR, CONSOLE_PROMPT);
-        font->render(renderer, edit_field_position, vec2(CONSOLE_FONT_SIZE, CONSOLE_FONT_SIZE),
-                     FONT_DEBUG_COLOR, String_View {edit_field_size, edit_field});
-
-        // CURSOR
-        const auto cursor_x = edit_field_cursor * BITMAP_FONT_CHAR_WIDTH * CONSOLE_FONT_SIZE + prompt_size.x;
-        {
-            SDL_Color cursor_color = FONT_DEBUG_COLOR;
-            const float blink_alpha = cosf(blink_angle * CONSOLE_BLINK_FREQUENCY) * 0.5f + 0.5f;
-            cursor_color.a = (Uint8) floorf(blink_alpha * 255.0f);
-            fill_rect(renderer,
-                      rect(vec2(cursor_x, edit_field_position.y),
-                           (float) (BITMAP_FONT_CHAR_WIDTH * CONSOLE_FONT_SIZE),
-                           (float) (BITMAP_FONT_CHAR_HEIGHT * CONSOLE_FONT_SIZE)),
-                      cursor_color);
-            if (edit_field_cursor < edit_field_size) {
-                SDL_Color overlay_text_color = CONSOLE_BACKGROUND_COLOR;
-                overlay_text_color.a = cursor_color.a;
-                font->render(renderer, vec2(cursor_x, edit_field_position.y), vec2(CONSOLE_FONT_SIZE, CONSOLE_FONT_SIZE),
-                             overlay_text_color, String_View {1, edit_field + edit_field_cursor});
-            }
-        }
-
-        // SELECTION
-        // text:              the q[uick bro]w[n fox jump]s over the lazy dog
-        // cursor:                           ^
-        // selection_begin:         ^                   ^
-        {
-            auto selection = get_selection();
-
-            if (!selection.is_empty()) {
-                const float begin_x = (float) selection.begin * BITMAP_FONT_CHAR_WIDTH * CONSOLE_FONT_SIZE + prompt_size.x;
-                const float end_x = (float) selection.end * BITMAP_FONT_CHAR_WIDTH * CONSOLE_FONT_SIZE + prompt_size.x;
-                const float width = end_x - begin_x;
-
-                fill_rect(renderer,
-                          rect(vec2(begin_x, edit_field_position.y),
-                               width,
-                               (float) (BITMAP_FONT_CHAR_HEIGHT * CONSOLE_FONT_SIZE)),
-                          CONSOLE_SELECTION_COLOR);
-
-                font->render(renderer, vec2(begin_x, edit_field_position.y), vec2(CONSOLE_FONT_SIZE, CONSOLE_FONT_SIZE),
-                             CONSOLE_BACKGROUND_COLOR,
-                             String_View {selection.end - selection.begin, edit_field + selection.begin});
-            }
-        }
+                 FONT_DEBUG_COLOR, CONSOLE_PROMPT);
+        edit_field.render(renderer, font, edit_field_position);
 
         // POPUP
         if (completion_popup_enabled) {
-            completion_popup.render(renderer, font, vec2(cursor_x, edit_field_position.y));
+            completion_popup.render(renderer, font, edit_field.cursor_position(edit_field_position));
         }
     }
 }
@@ -102,7 +42,7 @@ void Console::update(float dt)
 {
     if (enabled) {
         if (slide_position < 1.0f) slide_position += dt * CONSOLE_SLIDE_SPEED;
-        blink_angle = fmodf(blink_angle + 2 * PI * dt, 2 * PI);
+        edit_field.update(dt);
     } else {
         if (slide_position > 0.0f) slide_position -= dt * CONSOLE_SLIDE_SPEED;
     }
@@ -118,98 +58,10 @@ void Console::toggle()
     }
 }
 
-void Console::cursor_left(bool selection)
-{
-    if (edit_field_cursor > 0) {
-        edit_field_cursor -= 1;
-        if (!selection) {
-            edit_field_selection_begin = edit_field_cursor;
-        }
-    }
-}
-
-void Console::cursor_right(bool selection)
-{
-    if (edit_field_cursor < edit_field_size) {
-        edit_field_cursor += 1;
-        if (!selection) {
-            edit_field_selection_begin = edit_field_cursor;
-        }
-    }
-}
-
-void Console::insert_sv(String_View sv)
-{
-    const size_t remaining = CONSOLE_COLUMNS - edit_field_size;
-    const size_t n = min(sv.count, remaining);
-
-    memmove(edit_field + edit_field_cursor + n,
-            edit_field + edit_field_cursor,
-            edit_field_size - edit_field_cursor);
-    memcpy(edit_field + edit_field_cursor, sv.data, n);
-    edit_field_cursor += n;
-    edit_field_selection_begin = edit_field_cursor;
-    edit_field_size += n;
-}
-
-void Console::insert_cstr(const char *cstr)
-{
-    const size_t remaining = CONSOLE_COLUMNS - edit_field_size;
-    const size_t n = min(strlen(cstr), remaining);
-
-    memmove(edit_field + edit_field_cursor + n,
-            edit_field + edit_field_cursor,
-            edit_field_size - edit_field_cursor);
-    memcpy(edit_field + edit_field_cursor, cstr, n);
-    edit_field_cursor += n;
-    edit_field_selection_begin = edit_field_cursor;
-    edit_field_size += n;
-}
-
-void Console::delete_selection(Selection selection)
-{
-    assert(selection.begin < edit_field_size);
-    assert(selection.end <= edit_field_size);
-
-    memmove(edit_field + selection.begin,
-            edit_field + selection.end,
-            edit_field_size - selection.end);
-    edit_field_cursor = selection.begin;
-    edit_field_size -= selection.size();
-    edit_field_selection_begin = edit_field_cursor;
-}
-
-void Console::delete_char()
-{
-    if (edit_field_cursor < edit_field_size) {
-        const size_t n = edit_field_size - edit_field_cursor - 1;
-        memmove(
-            edit_field + edit_field_cursor,
-            edit_field + edit_field_cursor + 1,
-            n);
-        edit_field_size -= 1;
-        edit_field_selection_begin = edit_field_cursor;
-    }
-}
-
-void Console::backspace_char()
-{
-    if (edit_field_cursor > 0) {
-        const size_t n = edit_field_size - edit_field_cursor;
-        memmove(
-            edit_field + edit_field_cursor - 1,
-            edit_field + edit_field_cursor,
-            n);
-        edit_field_cursor -= 1;
-        edit_field_size -= 1;
-        edit_field_selection_begin = edit_field_cursor;
-    }
-}
-
 void Console::start_autocompletion()
 {
     completion_popup.clear();
-    String_View prefix = {edit_field_cursor, edit_field};
+    String_View prefix = {edit_field.edit_field_cursor, edit_field.edit_field};
     for (size_t i = 0; i < commands_count && !completion_popup.full(); ++i) {
         if (commands[i].name.has_prefix(prefix)) {
             completion_popup.push(commands[i].name);
@@ -236,142 +88,72 @@ void Console::handle_event(SDL_Event *event, Game *game)
                 } break;
                 case SDLK_RETURN: {
                     auto s = completion_popup.items[completion_popup.items_cursor];
-                    s.chop(edit_field_cursor);
-                    insert_sv(s);
+                    s.chop(edit_field.edit_field_cursor);
+                    edit_field.insert_sv(s);
                     completion_popup_enabled = false;
                 } break;
                 }
             } break;
             }
         } else {
-            blink_angle = 0.0;
-            switch (event->type) {
-            case SDL_KEYDOWN: {
-                switch (event->key.keysym.sym) {
-                case SDLK_SPACE: {
-                    if (event->key.keysym.mod & KMOD_LCTRL) {
-                        start_autocompletion();
-                    }
-                } break;
+            if (event->type == SDL_KEYDOWN && event->key.keysym.sym == SDLK_RETURN) {
+                scroll = 0;
+                String_View command_expr = edit_field.as_string_view().trim();
 
-                case SDLK_DELETE: {
-                    const auto selection = get_selection();
-                    if (selection.is_empty()) {
-                        delete_char();
-                    } else {
-                        delete_selection(selection);
-                    }
-                } break;
+                this->println(CONSOLE_PROMPT, edit_field.as_string_view());
+                history.push(edit_field.as_string_view());
+                edit_field.clean();
 
-                case SDLK_BACKSPACE: {
-                    const auto selection = get_selection();
-                    if (selection.is_empty()) {
-                        backspace_char();
-                    } else {
-                        delete_selection(selection);
-                    }
-                } break;
-
-                case SDLK_LEFT: {
-                    cursor_left(event->key.keysym.mod & KMOD_LSHIFT);
-                } break;
-
-                case SDLK_RIGHT: {
-                    cursor_right(event->key.keysym.mod & KMOD_LSHIFT);
-                } break;
-
-                case SDLK_UP: {
-                    const auto j = history.current();
-                    if (j >= 0) {
-                        memcpy(edit_field, &history.entries[j], history.entry_sizes[j]);
-                        edit_field_size = history.entry_sizes[j];
-                        edit_field_cursor = history.entry_sizes[j];
-                        edit_field_selection_begin = edit_field_cursor;
-                    }
-                    history.up();
-                } break;
-
-                case SDLK_DOWN: {
-                    const auto j = history.current();
-                    if (j >= 0) {
-                        memcpy(edit_field, &history.entries[j], history.entry_sizes[j]);
-                        edit_field_size = history.entry_sizes[j];
-                        edit_field_cursor = history.entry_sizes[j];
-                        edit_field_selection_begin = edit_field_cursor;
-                    }
-                    history.down();
-                } break;
-
-                case SDLK_v: {
-                    if (event->key.keysym.mod & KMOD_LCTRL) {
-                        insert_cstr(sec(SDL_GetClipboardText()));
-                    }
-                } break;
-
-                case SDLK_c: {
-                    if (event->key.keysym.mod & KMOD_LCTRL) {
-                        auto selection = get_selection();
-                        if (!selection.is_empty()) {
-                            memcpy(clipboard_buffer, edit_field + selection.begin, selection.size());
-                            clipboard_buffer[selection.size()] = '\0';
-                            sec(SDL_SetClipboardText(clipboard_buffer));
+                if (command_expr.count > 0) {
+                    bool command_found = false;
+                    const auto command_name = command_expr.chop_word();
+                    for (size_t i = 0; !command_found && i < commands_count; ++i) {
+                        if (commands[i].name == command_name) {
+                            commands[i].run(game, command_expr);
+                            command_found = true;
                         }
                     }
-                } break;
 
-                case SDLK_RETURN: {
-                    scroll = 0;
-                    String_View command_expr = String_View {edit_field_size, edit_field}.trim();
-
-                    this->println(CONSOLE_PROMPT, String_View {edit_field_size, edit_field});
-                    history.push(edit_field, edit_field_size);
-                    edit_field_size = 0;
-                    edit_field_cursor = 0;
-                    edit_field_selection_begin = 0;
-
-                    if (command_expr.count > 0) {
-                        bool command_found = false;
-                        const auto command_name = command_expr.chop_word();
-                        for (size_t i = 0; !command_found && i < commands_count; ++i) {
-                            if (commands[i].name == command_name) {
-                                commands[i].run(game, command_expr);
-                                command_found = true;
-                            }
-                        }
-
-                        if (!command_found) {
-                            this->println("Command `", command_name, "` not found");
-                        }
+                    if (!command_found) {
+                        this->println("Command `", command_name, "` not found");
                     }
-                } break;
                 }
-            } break;
-
-            case SDL_MOUSEWHEEL: {
+            } else if (event->type == SDL_KEYDOWN && event->key.keysym.sym == SDLK_SPACE && event->key.keysym.mod & KMOD_LCTRL) {
+                start_autocompletion();
+            } else if (event->type == SDL_KEYDOWN && event->key.keysym.sym == SDLK_UP) {
+                const auto j = history.current();
+                if (j >= 0) {
+                    edit_field.copy_from_string_view(String_View {history.entry_sizes[j], history.entries[j]});
+                }
+                history.up();
+            } else if (event->type == SDL_KEYDOWN && event->key.keysym.sym == SDLK_DOWN) {
+                const auto j = history.current();
+                if (j >= 0) {
+                    edit_field.copy_from_string_view(String_View {history.entry_sizes[j], history.entries[j]});
+                }
+                history.down();
+            } else if (event->type == SDL_MOUSEWHEEL ) {
                 if (event->wheel.y > 0) {
                     scroll += 1;
                 } else if (event->wheel.y < 0) {
                     if (scroll > 0) scroll -= 1;
                 }
-            } break;
-
-            case SDL_TEXTINPUT: {
-                insert_cstr(event->text.text);
-            } break;
+            } else {
+                edit_field.handle_event(event);
             }
         }
     }
 }
 
-void Console::History::push(char *entry, size_t entry_size)
+void Console::History::push(String_View entry)
 {
-    if (entry_size > CONSOLE_COLUMNS) {
-        entry_size = CONSOLE_COLUMNS;
+    if (entry.count > CONSOLE_COLUMNS) {
+        entry.count = CONSOLE_COLUMNS;
     }
 
     const size_t j = (begin + count) % CONSOLE_HISTORY_CAPACITY;
-    memcpy(&entries[j], entry, entry_size);
-    entry_sizes[j] = entry_size;
+    memcpy(&entries[j], entry.data, entry.count);
+    entry_sizes[j] = entry.count;
 
     if (count < CONSOLE_HISTORY_CAPACITY) {
         count += 1;
